requirements.txt
# Core dependencies
python-dotenv>=0.19.0
requests>=2.31.0
GitPython>=3.1.40
chardet>=5.1.0
tiktoken>=0.5.1
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-community>=0.0.10
pinecone-client[grpc]>=3.0.0
protobuf>=4.25.1

# Data processing
numpy>=1.24.0
pandas>=2.0.0

# Text processing
nltk>=3.8.1
spacy>=3.7.0
langchain-text-splitters>=0.0.1

# Utilities
tqdm>=4.65.0
python-dateutil>=2.8.2

# Development tools
black>=23.3.0
flake8>=6.0.0
pytest>=7.3.1

# Optional dependencies
pydantic>=2.5.0